Fixing wrong calculation for Cyrillic strings
The .size() function for string variables returns the length of the string, in terms of bytes (http://www.cplusplus.com/reference/string/string/size/). Not the number of characters. So, for Cyrillic strings, this numbers are different. This leads to the wrong result. You can check <select EditDistance('йцфы', '');> It will return 8, instead of 4.
The solution is to use wstring type.